==> 01_Ingest_csv_to_sagemakers.ipynb <==
## Ingest CSV files to Sagemaker


import pandas as pd
import boto3
import sagemaker
import matplotlib.pyplot as plt
import os
import numpy as np

# Locate S3 bucket we created manually

!aws s3 ls s3://projectads508/aisles/

!aws s3 ls s3://projectads508/departments/

!aws s3 ls s3://projectads508/prior/

!aws s3 ls s3://projectads508/train/

!aws s3 ls s3://projectads508/orders/

!aws s3 ls s3://projectads508/products/


sess= sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
account_id = boto3.client("sts").get_caller_identity().get("Account")

sm = boto3.Session().client(service_name="sagemaker", region_name=region)

# Set S3 Source Location (Public S3 Bucket)

s3_public_path_csv = "s3://projectads508"
%store s3_public_path_csv

# Set S3 Destination Location (Our Private S3 Bucket)

s3_private_path_csv = "s3://{}/projectads508".format(bucket)
print(s3_private_path_csv)

# Copy Data From the Public S3 Bucket to our Private S3 Bucket in this Account


!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "aisles/aisles.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "departments/departments.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "prior/order_products__prior.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "train/order_products__train.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "orders/orders.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "products/products.csv"

print(s3_private_path_csv)

!aws s3 ls $s3_private_path_csv/aisles/

!aws s3 ls $s3_private_path_csv/departments/

!aws s3 ls $s3_private_path_csv/prior/

!aws s3 ls $s3_private_path_csv/train/

!aws s3 ls $s3_private_path_csv/orders/

!aws s3 ls $s3_private_path_csv/products/

from IPython.core.display import display, HTML

display(
    HTML(
        '<b>Review <a target="blank" href="https://s3.console.aws.amazon.com/s3/buckets/sagemaker-{}-{}/projectads508/?region={}&tab=overview">S3 Bucket</a></b>'.format(
            region, account_id, region
        )
    )
)
