==> 02_Athena_database.ipynb <==
## Create Database in Athena


import pandas as pd
import boto3
import sagemaker
import matplotlib.pyplot as plt
import os
import numpy as np

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

!pip install --disable-pip-version-check -q PyAthena==2.1.0
from pyathena import connect

database_name = "oneclickaws"
%store database_name

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://{0}/athena/staging".format(bucket)

conn = connect(region_name = region, s3_staging_dir=s3_staging_dir)

statement = "CREATE DATABASE IF NOT EXISTS {}".format(database_name)
print(statement)

pd.read_sql(statement, conn)

# Verify The Database Has Been Created Succesfully

statement = "SHOW DATABASES"

df_show = pd.read_sql(statement, conn)
df_show.head(5)

if database_name in df_show.values:
    ingest_create_athena_db_passed = True

%store ingest_create_athena_db_passed

%store
