==> 03_Register_csv_to_athena .ipynb <==
import pandas as pd
import boto3
import sagemaker

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

from pyathena import connect

ingest_create_athena_table_csv_passed = False

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://{0}/athena/staging".format(bucket)



%store -r ingest_create_athena_db_passed


## Creating 6 tables in Athena

s3_private_path_csv1 = "s3://{}/projectads508/aisles/".format(bucket)
print(s3_private_path_csv1)

s3_private_path_csv2 = "s3://{}/projectads508/departments/".format(bucket)
print(s3_private_path_csv2)

s3_private_path_csv3 = "s3://{}/projectads508/prior/".format(bucket)
print(s3_private_path_csv3)

s3_private_path_csv4 = "s3://{}/projectads508/train/".format(bucket)
print(s3_private_path_csv4)

s3_private_path_csv5 = "s3://{}/projectads508/orders/".format(bucket)
print(s3_private_path_csv5)

s3_private_path_csv6 = "s3://{}/projectads508/products/".format(bucket)
print(s3_private_path_csv6)

# Set Athena parameters
table1 = "aisles"
table2= "departments"
table3= "order_products_prior"
table4= "order_products_train"
table5= "orders"
table6= "products"
database_name='oneclickaws'

conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

# SQL statement to execute aisles table
statement1 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         aisle_id int,
         aisle string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table1, s3_private_path_csv1
)


pd.read_sql(statement1, conn)

# SQL statement to execute department table
statement2 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         department_id int,
         department string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table2, s3_private_path_csv2
)


pd.read_sql(statement2, conn)

# SQL statement to execute order_products__prior table
statement3 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         product_id int,
         add_to_cart_order int,
         reordered string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table3, s3_private_path_csv3
)


pd.read_sql(statement3, conn)

# SQL statement to execute order_products__train table
statement4 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         product_id int,
         add_to_cart_order int,
         reordered string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table4, s3_private_path_csv4
)


pd.read_sql(statement4, conn)

# SQL statement to execute orders table
statement5 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         user_id int,
         eval_set string,
         order_number int,
         order_dow string,
         order_hour_of_day string,
         days_since_prior_order float
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table5, s3_private_path_csv5
)


pd.read_sql(statement5, conn)

# SQL statement to execute products table
statement6= """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
          product_id int,
         product_name string,
         aisle_id int,
         department_id int
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table6, s3_private_path_csv6
)


pd.read_sql(statement6, conn)

## Test created database and tables

statement = "SHOW TABLES in {}".format(database_name)

df_show = pd.read_sql(statement, conn)
df_show.head(6)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table1
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table2
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

test_statement = """SELECT * FROM {}.{}  order by order_id LIMIT 12  """.format(
    database_name, table3
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(12)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table4
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

test_statement = """SELECT * FROM {}.{} order by user_id LIMIT 12 """.format(
    database_name, table5
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(12)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table6
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

%store
