==> 09 _Data_Preparation.ipynb <==

## Data Preparation

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import gc
pd.options.mode.chained_assignment = None

## Reading all data

orders = pd.read_csv("./data-clarify/orders.csv")
order_products_train = pd.read_csv('./data-clarify/order_products__train.csv')
aisles = pd.read_csv('./data-clarify/aisles.csv')
products = pd.read_csv('./data-clarify/products.csv')
departments = pd.read_csv('./data-clarify/departments.csv')
product_features = pd.read_pickle('./data-clarify/product_features.pkl')
user_features = pd.read_pickle('./data-clarify/user_features.pkl')

## Merging train order data with orders

train_orders = orders.merge(order_products_train, on = 'order_id', how = 'inner')
train_orders.head()

removing unnecessary columns from train_orders

train_orders.drop(['eval_set', 'add_to_cart_order', 'order_id'], axis = 1, inplace = True)

train_orders.shape

## Merging with product_features and user_features

df = train_orders.merge(product_features, on = 'product_id', how = 'left')
df = df.merge(user_features, on = 'user_id', how = 'left')
df.head()


df.shape

df.columns

path = "./data-clarify/final_data.csv"

df.to_csv(path)

df['reordered'].value_counts()

## Create dummy variables
Convert Day and Hour categorical variables to dummy variables.

final_data = pd.get_dummies(data=df, prefix=['Day','Hour'], columns=['order_dow','order_hour_of_day'], drop_first=True)
final_data.head()

### Duplicate rows
 The dataset doesn't have any duplicates.

    
counter = 0
r,c = final_data.shape

df1 = final_data.drop_duplicates()
df1.reset_index(drop=True,inplace=True)

if df1.shape==(r,c):
    print('\n\033[1mInference:\033[0m The dataset doesn\'t have any duplicates')
else:
    print(f'\n\033[1mInference:\033[0m Number of duplicates dropped ---> {r-df.shape[0]}')

## Save the new dataset for future 
save the dataset with dummy variables for future tasks.

!pip install --disable-pip-version-check -q PyAthena==2.1.0
from pyathena import connect
import sagemaker
import boto3
sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
from pyathena import connect


path = "./data-clarify/final_data.csv"
final_data.to_csv(path, index=False, header=True)

import time

timestamp = int(time.time())

dummy_data_s3_uri = sess.upload_data(bucket=bucket, key_prefix="dummy-dataset-{}".format(timestamp), path=path)
dummy_data_s3_uri

!aws s3 ls $dummy_data_s3_uri

%store dummy_data_s3_uri

%store
