==> 10_Autopilot.ipynb <==
## Prepare data for Autopilot Sagemaker

import pandas as pd
import numpy as np

import sagemaker
import boto3

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

sm = boto3.Session().client(service_name="sagemaker", region_name=region)
s3 = boto3.Session().client(service_name="s3", region_name=region)

 ## take a look at the data

data = pd.read_csv("./data-clarify/final_data.csv")
pd.set_option('display.max_columns',500)
pd.set_option('display.max_rows',10)
data

data.shape

## Import the ‘train’ and ‘validation’ datasets

train_data=data.sample(frac=0.8,random_state=200)
test_data=data.drop(train_data.index)
test_data_no_target=test_data.drop(columns=['reordered'])

## upload the data to S3

train_file='train_data.csv';
train_data.to_csv(train_file,index=False, header=True)
train_data_s3_path=sess.upload_data(path=train_file)
print('Train data uploaded to:' + train_data_s3_path)

test_file='test_data.csv';
test_data_no_target.to_csv(test_file,index=False, header=False)
test_data_s3_path=sess.upload_data(path=test_file)
print('Test data uploaded to:' + test_data_s3_path)
