==> 15_xgboost.ipynb <==
import sagemaker
import boto3
from sagemaker.session import s3_input, Session


sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

sm = boto3.Session().client(service_name="sagemaker", region_name=region)
s3 = boto3.Session().client(service_name="s3", region_name=region)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import gc
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv('./data-clarify/final_data.csv')


training_data, testing_data = train_test_split(df, test_size=0.2, random_state=25)


training_data 

# Target variable must be in 1st column and data should be in csv format for XG boost

col_name="reordered"
first_col = training_data.pop(col_name)

training_data.insert(0, col_name, first_col)

# Header must be removed

training_data.columns = range(training_data.shape[1])
print (training_data)

training_data.head(5)

path_train = "./data-clarify/train_data.csv"
training_data.to_csv(path_train, index=False, header=True)

import time

timestamp = int(time.time())

train_data_s3_uri = sess.upload_data(bucket=bucket, key_prefix="traindata-{}".format(timestamp), path=path_train)
train_data_s3_uri

path_test = "./data-clarify/test_data.csv"
testing_data.to_csv(path_test, index=False, header=True)

import time

timestamp = int(time.time())

test_data_s3_uri = sess.upload_data(bucket=bucket, key_prefix="testdata-{}".format(timestamp), path=path_test)
test_data_s3_uri

container = sagemaker.image_uris.retrieve(region= boto3.Session().region_name, framework= 'xgboost', version = 'latest')

prefix = 'xgboostframework'

xgb= sagemaker.estimator.Estimator(container, role, train_instance_count =1, train_instance_type = 'ml.m4.xlarge', output_path = 's3://{}/{}/output'.format(bucket,prefix), sagemaker_session =sess)

xgb.set_hyperparameters(max_depth= 5,
        eta=0.2,
        gamma=4,
        min_child_weight=6,
        subsample=0.8,
        objective= 'binary:logistic',
        num_round=100)

content_type = "csv"
train_input = s3_input(train_data_s3_uri, content_type=content_type)
test_input = s3_input(test_data_s3_uri, content_type=content_type)





model= xgb.fit({'train':train_input})
